requirements.txt
amqp==5.0.2
aniso8601==7.0.0
asgiref==3.2.7
autopep8==1.5.4
billiard==3.6.3.0
boto3==1.16.51
botocore==1.19.51
cachetools==4.1.1
celery==5.0.5
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-cors-headers==3.5.0
django-filter==2.4.0
django-ses==1.0.3
django-storages==1.11.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
facebook-sdk==3.1.0
Faker==4.1.0
future==0.18.2
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
Pillow==8.1.0
promise==2.3
prompt-toolkit==3.0.10
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycurl==7.43.0.5
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
python-twitter==3.5
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Rx==1.6.1
s3transfer==0.3.3
singledispatch==3.4.0.3
six==1.15.0
slugify==0.0.1
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
twitter==1.18.0
uritemplate==3.0.1
urllib3==1.25.9
vine==5.0.0
wcwidth==0.2.5
